It was decided to delete the unique_token.rb file, add that code to the link and add the uniqueness verification in link_validations.rb
